feat: Add function to rotate a 2D matrix 90 degrees clockwise
This commit adds a new function called `rotate_2d_matrix` that can be used to rotate a 2D matrix 90 degrees clockwise. The function takes a matrix as input and modifies it in-place. It uses a nested loop to swap the elements in the matrix in a clockwise direction. This function will be useful for tasks that require rotating a matrix, such as image processing or game development.

Note: This commit message does not include any meta information like issue references, tags, or author names.